- cleaned up the video scale mode code a bit, in order to increase future flexibility.
